+++ changelog.md
@@ -3,6 +3,9 @@ Naia - Project Ethea Phase 1 Shared Library - Changelog
 
 Version 20210430+dev:
 ---------------------
+* Units:
+  * Verlissh Matrix Core and Verlissh Matrix Flow System can now have randomly
+    generated names in campaigns that define `ENABLE_VERLISSH_MACHINE_NAMES`.
 
 
 Version 20210430: